fix loading in createResponse
